pr fixes cherry picked from commit: ""pr fixes 2ae189b"
