+++ KarhooUISDK/Screens/CheckoutScreen/Checkout/VehicleDetailsCard.swift
@@ -75,39 +75,3 @@ struct VehicleDetailsCardViewModel {
     let fleetIconUrl: String
     let cancelationText: String?
 }
-
-struct NewVehicleCapacity: View {
-    var passangerCapacity: Int
-    var luggageCapacity: Int
-    var body: some View {
-        HStack {
-            CapacityCard(iconName: "kh_uisdk_passenger_capacity_icon", value: passangerCapacity)
-            CapacityCard(iconName: "kh_uisdk_luggage_icon", value: luggageCapacity)
-        }
-    }
-    
-    struct CapacityCard: View {
-        var iconName: String
-        var value: Int
-        var body: some View {
-            HStack {
-                Group {
-                    Image(uiImage: .uisdkImage(iconName))
-                        .frame(
-                            width: UIConstants.Dimension.Icon.small,
-                            height: UIConstants.Dimension.Icon.small
-                        )
-                    Text("\(value)")
-                        .font(Font(KarhooUI.fonts.footnoteSemiold()))
-                        .foregroundColor(Color(KarhooUI.colors.text))
-                }
-            }
-            .padding(UIConstants.Spacing.xSmall)
-            .background(
-                RoundedRectangle(cornerRadius: UIConstants.CornerRadius.xSmall)
-                    .fill(Color(KarhooUI.colors.background1))
-            )
-        }
-    }
-    
-}
